app/workflow/workflow_engine.py
from typing import List, Dict, Any
from app.knowledge.knowledge_graph import KnowledgeGraph
from app.recommendation.recommendation_system import RecommendationSystem
from app.agents.factory import AgentFactory
import logging
import asyncio

logger = logging.getLogger(__name__)

class WorkflowEngine:
    def __init__(self, knowledge_graph: KnowledgeGraph, recommendation_system: RecommendationSystem, agent_factory: AgentFactory):
        self.knowledge_graph = knowledge_graph
        self.recommendation_system = recommendation_system
        self.agent_factory = agent_factory

    async def create_workflow(self, task_analysis: Dict[str, Any]) -> List[Dict[str, Any]]:
        # Implementation
        return [{"step": "Step 1"}, {"step": "Step 2"}]  # Placeholder implementation

    async def execute_workflow(self, workflow: List[str]) -> Dict[str, Any]:
        results = []
        for step in workflow:
            if asyncio.iscoroutinefunction(self.agent_factory.create_agent):
                agent = await self.agent_factory.create_agent({"content": step})
            else:
                agent = self.agent_factory.create_agent({"content": step})
            result = await agent.process_task({"content": step})
            results.append(result)
        return {"results": results}

    async def _optimize_workflow(self, workflow: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        # Implementation
        pass